LPS-105162 move 1_1_3 to 1_1_4
